# openapi/openapi.json
{
  "openapi": "3.0.3",
  "info": { "title": "X07 Registry API", "version": "0.1.0" },
  "paths": {
    "/healthz": {
      "get": {
        "responses": {
          "200": { "description": "OK" }
        }
      }
    },
    "/index/config.json": {
      "get": {
        "responses": {
          "200": { "description": "Index config" }
        }
      }
    },
    "/index/{path}": {
      "get": {
        "parameters": [
          { "name": "path", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Index file contents" },
          "404": { "description": "Not found" }
        }
      }
    },
    "/v1/auth/token": {
      "post": {
        "responses": {
          "200": { "description": "Token is valid" },
          "401": { "description": "Unauthorized" }
        }
      }
    },
    "/v1/packages/publish": {
      "post": {
        "responses": {
          "200": { "description": "Published" },
          "401": { "description": "Unauthorized" },
          "409": { "description": "Already published" }
        }
      }
    },
    "/v1/packages/{name}/{version}/download": {
      "get": {
        "parameters": [
          { "name": "name", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "version", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Package archive bytes" },
          "404": { "description": "Not found" }
        }
      }
    },
    "/v1/packages/{name}/{version}/metadata": {
      "get": {
        "parameters": [
          { "name": "name", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "version", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Package metadata" },
          "404": { "description": "Not found" }
        }
      }
    }
  }
}